gh-131238: Add pycore_interpframe_structs.h header
Add an explicit include yo pycore_interpframe_structs.h in
pycore_runtime_structs.h to fix a dependency cycle.
