+++ Include/internal/pycore_genobject.h
@@ -9,41 +9,8 @@ extern "C" {
 #endif
 
 #include "pycore_interpframe.h"   // _PyInterpreterFrame
+#include "pycore_interpframe_structs.h" // _PyGenObject
 
-
-/* _PyGenObject_HEAD defines the initial segment of generator
-   and coroutine objects. */
-#define _PyGenObject_HEAD(prefix)                                           \
-    PyObject_HEAD                                                           \
-    /* List of weak reference. */                                           \
-    PyObject *prefix##_weakreflist;                                         \
-    /* Name of the generator. */                                            \
-    PyObject *prefix##_name;                                                \
-    /* Qualified name of the generator. */                                  \
-    PyObject *prefix##_qualname;                                            \
-    _PyErr_StackItem prefix##_exc_state;                                    \
-    PyObject *prefix##_origin_or_finalizer;                                 \
-    char prefix##_hooks_inited;                                             \
-    char prefix##_closed;                                                   \
-    char prefix##_running_async;                                            \
-    /* The frame */                                                         \
-    int8_t prefix##_frame_state;                                            \
-    _PyInterpreterFrame prefix##_iframe;                                    \
-
-struct _PyGenObject {
-    /* The gi_ prefix is intended to remind of generator-iterator. */
-    _PyGenObject_HEAD(gi)
-};
-
-struct _PyCoroObject {
-    _PyGenObject_HEAD(cr)
-};
-
-struct _PyAsyncGenObject {
-    _PyGenObject_HEAD(ag)
-};
-
-#undef _PyGenObject_HEAD
 
 static inline
 PyGenObject *_PyGen_GetGeneratorFromFrame(_PyInterpreterFrame *frame)
